requirements.txt
Django==2.0.7
djangorestframework==3.8.2
markdown==2.6.11
django-filter==2.0.0
django-cleanup==2.1.0
Pillow==5.2.0
django-cors-headers==2.4.0
openpyxl==2.5.5
requests